- Small release build datatype warning fix
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1996 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/ChatHandler.cpp
@@ -688,7 +688,7 @@ void ChatHandler::handleChat(Player *player, PacketReader &packet) {
 						player->getInventory()->setMesos(atoi(args.c_str()));
 					break;
 				case CmdRelog:
-					player->changeChannel(ChannelServer::Instance()->getChannel());
+					player->changeChannel((int8_t)ChannelServer::Instance()->getChannel());
 					break;
 				case CmdSave:
 					player->saveAll();
